pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "conda-forge-converter"
dynamic = ["version"] # Version is dynamically determined from _version.py
description = "Tools for converting Anaconda environments to conda-forge and validating GitHub Actions workflows"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Adam Rice", email = "rice.adam@gmail.com" }]
maintainers = [{ name = "Adam Rice", email = "rice.adam@gmail.com" }]
keywords = [
    "conda",
    "conda-forge",
    "anaconda",
    "environment",
    "virtual environment",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pyyaml>=6.0.1",
    "typing_extensions>=4.10.0", # For additional type hints if needed
    "click>=8.1.0",              # Modern CLI framework
    "rich>=13.0.0",              # Terminal formatting
    "pydantic>=2.0.0",           # Data validation and settings management
]
[project.optional-dependencies]
validation = ["actionlint>=1.6.0"]
dev = [
    "ruff>=0.3.0",
    "pyright>=1.1.350",
    "python-semantic-release>=9.0.0",
    "pre-commit>=3.5.0",
    "build>=1.0.0",
    "hatch>=1.9.0",
    "types-pyyaml>=6.0.12.20250402",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.3.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python>=1.8.0",
    "mkdocs-minify-plugin>=0.7.0",
    "mkdocs-redirects>=1.2.1",
    "hatch>=1.9.0",
]
all = [
    # Include all dependencies from other groups directly
    "actionlint>=1.6.0", # validation
    # dev dependencies
    "ruff>=0.3.0",
    "pyright>=1.1.350",
    "python-semantic-release>=9.0.0",
    "pre-commit>=3.5.0",
    "build>=1.0.0",
    "hatch>=1.9.0",
    "types-pyyaml>=6.0.12.20250402",
    # test dependencies
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.3.0",
    # docs dependencies
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings>=0.24.0",
    "mkdocstrings-python>=1.8.0",
    "mkdocs-minify-plugin>=0.7.0",
    "mkdocs-redirects>=1.2.1",
]
[project.scripts]
anaconda-to-miniconda = "conda_forge_converter.cli:main"
validate-github-workflows = "github_actions_validator.cli:main"
validate-github-workflows-standard = "github_actions_validator.cli_standard:main"


[project.urls]
Homepage = "https://github.com/yourusername/conda-forge-converter"
Documentation = "https://github.com/yourusername/conda-forge-converter#readme"
Issues = "https://github.com/yourusername/conda-forge-converter/issues"
Source = "https://github.com/yourusername/conda-forge-converter"

# Configure dynamic version management - using python-semantic-release
[tool.hatch.version]
source = "regex"
regex = "__version__ = '(?P<version>[^']+)'"
path = "src/conda_forge_converter/_version.py"

# Also track github_actions_validator version
[tool.hatch.version.github_actions_validator]
source = "regex"
regex = "__version__ = '(?P<version>[^']+)'"
path = "src/github_actions_validator/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/conda_forge_converter", "src/github_actions_validator"]

# Environment management is now handled by UV
[tool.hatch.envs.default.scripts]
bump-version = "semantic-release version --no-commit --no-tag --no-push --no-vcs-release"
prepare-release = "semantic-release version --print"
check-next-version = "semantic-release version --print"

# Test scripts - using pytest.ini and .coveragerc for configuration
test = "pytest"
test-unit = "pytest -m 'unit'"
test-integration = "pytest -m 'integration'"
test-github-actions = "pytest tests/test_github_actions_validator.py tests/test_coverage.py"
test-conda-forge = "pytest tests/test_cli.py"
test-fast = "pytest -k 'not slow and not integration'"
test-specific = "pytest"
clean-coverage = "python -c \"import pathlib; [p.unlink() for p in pathlib.Path('.').glob('.coverage*') if p.is_file()]; [p.unlink() for p in pathlib.Path('.').glob('coverage.xml') if p.is_file()]; import shutil; shutil.rmtree('htmlcov', ignore_errors=True)\""
update-coverage-threshold = "python scripts/update_coverage_threshold.py"

# Linting and type checking
lint = "ruff check ."
format = "ruff format"
type-check = "pyright"

# Documentation scripts
docs-build = "mkdocs build --strict"
docs-serve = "mkdocs serve"

# Type checking configuration is now in pyrightconfig.json

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PT", "RUF", "C90", "D", "ANN"]
ignore = ["D203", "D213"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = false
mypy-init-return = true
suppress-none-returning = true
suppress-dummy-args = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"tests/*" = ["E501", "ANN", "PT009", "D"]
"scripts/*" = ["E501", "ANN201", "D"]
".github/scripts/*" = ["E501", "ANN", "C901", "D"]
# Ignore missing return type annotations in decorator wrappers
"src/conda_forge_converter/progress.py" = ["ANN401"]
"src/conda_forge_converter/cli.py" = ["E501"]
"src/github_actions_validator/cli.py" = ["E501", "C901"]
"src/github_actions_validator/runners/act.py" = ["C901"]
"src/github_actions_validator/validators/act.py" = ["C901"]
"src/github_actions_validator/validators/actionlint.py" = ["C901"]
"src/github_actions_validator/validators/execution.py" = ["C901"]
"src/github_actions_validator/validators/syntax.py" = ["C901"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# Pytest configuration is now in pytest.ini

# Python-Semantic-Release configuration
[tool.semantic_release]
version_variable = [
    "src/conda_forge_converter/_version.py:__version__",
    "src/github_actions_validator/_version.py:__version__",
]
branch = "master"
upload_to_pypi = true
upload_to_release = true
build_command = "python -m build"
github_token = "GH_TOKEN"

# Branch-specific configuration
major_on_zero = false     # Don't treat 0.x versions specially
tag_format = "v{version}"

# Commit message parsing
commit_parser = "conventional"
commit_author = "github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>"

# Changelog configuration
changelog_file = "CHANGELOG.md"
changelog_sections = [
    "feature",
    "fix",
    "breaking",
    "documentation",
    "performance",
    "refactor",
]
changelog_scope_section_order = [
    "core",
    "cli",
    "utils",
    "deps",
    "tests",
    "docs",
]

# Version management for development branches
[tool.semantic_release.branches.develop]
match = "develop"
prerelease = true
prerelease_token = "dev"

# Version management for feature branches
[tool.semantic_release.branches.feature]
match = "feature/*"
prerelease = true
prerelease_token = "feature"

# UV configuration for dependency management
[tool.uv]
# UV will automatically use requirements.lock for reproducible installations

# Coverage configuration is now in .coveragerc

[dependency-groups]
dev = ["types-pyyaml>=6.0.12.20250402"]
